Use text as the key for the (value, text) pairs.
Selectorablium acts as a wrapper for a select with value, text options.
The internals of the implementation use value as a unique key and this
forbids multiple texts with the same value which is a common scenario
in web pages. This commit changes the internals to use text as the key.
This way of course you can't have the same key text with different value
but this is a very rare case