Add support for cache invalidation via cache_control_url
This commit adds support for local cache invalidation, by accepting
a cache_control_url param. This URL should correspond to an endpoint
returning a single "last-modified-at" timestamp for the respective
resource collection.

If given, the StorageFreak submodule will make an XHR request to
that URL and depending on the timestamp response, it will issue a
db refresh request or not.

Note that this check is below the pre-existing `localCacheTimeout`
check, so if that is exceeded the db will get updated without
checking for last modification time.