Add support for conditional GET cache revalidation
This commit adds support for revalidating the local storage cache,
using the If-Modified-Since/Last-Modified HTTP headers.

In particular:

* It stores the last modified timestamp, if the Last-Modified header
is set by the remote server. The timestamp is stored in local storage.
* If a last modified timestamp is stored, it sends the If-Modified-Since
header when requesting remote data. This happens only if there is no
query, as we don't want to cache partial data.
* If the remote server responds to the above request with 304 Not Modified,
it just reads data available in the local storage cache.
* If the `cache_revalidate` option is enabled, it performs the above
"conditional GET" request at init time.