Fix evergreen specs for remote search
For these specs, the `done()` callback was being called before an
expectation was registered, so they weren't actually testing anything.

The reason for that was that the "dbsearch_results" event is triggered
twice, first for local and then for remote search, but the `done`
callback was being called in between.

Also, the following modifications were needed:

* Had to restore fake clock methods before `@respondJSON` to avoid
interfering with jQuery $.ajax internals when resolving promises
* Had to wrap async expectations in try/catch blocks so test errors
would be correctly displayed, instead of timing out